pyproject.toml
[project]
name = "otomopy"
version = "1.0.0"
dependencies = [
    "discord.py >= 2.0.0",
    "python-dotenv >= 0.19.0",
    "aiohttp >= 3.8.0",
]
requires-python = ">=3.11"
description = "Holodex relay bot for Discord"
license = "MIT"
license-files = ["LICENSE"]

[project.scripts]
otomopy = "otomopy.bot:main"
